Add default license, and icon file
