+++ scripts/src/lib/template/git.ts
@@ -1,0 +1,11 @@
+import type { ComputedConfiguration, TemplateWriter } from "./template";
+
+import gitignore from './templates/git/gitignore?raw';
+import workflow from './templates/git/workflow.yml?raw';
+import license from './templates/git/LICENSE?raw';
+
+export async function addGitFiles(writer: TemplateWriter, config: ComputedConfiguration) {
+    await writer.write('.gitignore', gitignore);
+    await writer.write('.github/workflows/build.yml', workflow);
+    await writer.write('LICENSE', license);
+}
