+++ scripts/src/lib/template/modjson.ts
@@ -2,6 +2,7 @@ import type { ComputedConfiguration, TemplateWriter } from "./template";
 import { generateMixin } from "./mixin";
 import { generateEntrypoint } from "./modentrypoint";
 import { getJavaVersion, getMinorMinecraftVersion } from "./java"
+import { decode64 } from './utils';
 
 export async function addModJson(writer: TemplateWriter, config: ComputedConfiguration) {
   var fabricModJson : any = {
@@ -17,6 +18,8 @@ export async function addModJson(writer: TemplateWriter, config: ComputedConfigu
       "homepage": "https://fabricmc.net/",
       "sources": "https://github.com/FabricMC/fabric-example-mod"
     },
+    "license": "CC0-1.0",
+    "icon": `assets/${config.modid}/icon.png`,
     "environment": "*",
     "entrypoints": await generateEntrypoint(writer, config),
     "mixins": await generateMixin(writer, config),
@@ -40,4 +43,8 @@ export async function addModJson(writer: TemplateWriter, config: ComputedConfigu
   }
 
   await writer.write("src/main/resources/fabric.mod.json", JSON.stringify(fabricModJson, null, "\t"));
+  await writer.write(`src/main/resources/assets/${config.modid}/icon.png`, decode64(ICON));
 }
+
+// Todo can we generate an icon from the mod name?
+const ICON = "iVBORw0KGgoAAAANSUhEUgAAAIAAAACAAQMAAAD58POIAAAABlBMVEUAAAD///+l2Z/dAAABeklEQVRIx9XTsW1cMQwGYAoKIlfWbaAVUroKbxSPcBtIQRaTN9EILFkI+l3ovXuSLvYZCK4wK+IrWBD/T1iGHg41fBUcygwWKY7QLGiCuoLaNoOY+1BnKFfQDUhnyGeZIZ3lNEO+LFBef01gyutlAlvOEzRb6MIDVCskDAD01MEJCTeKjWgHI1wp1A3Ui9Wg5HSHoFaDkN1BgroOzwN4ORkJuQNXX738NsKFAULpEI2wdIhAdRLTCM1JzBIHsMJZokSAkK/ApQMAbMADhCycDzASsoQObwDwR31Sn154AFJHpwPqT6NENACZOgIambYBNrCYAOSQflAcIEDCmPXEDyrQ42GP9wBuBXuAMqBENxBneCGegVNYIPsFygrqFqh2gXYDZoG+j5BuIC6QVyh8D444aPgXBKi/B9XtUD+ANkH9CsAuNz4D/wH8/Rx4gPZIsP8DDlAeHrRBXUFXkCugQ/YLUJhBiCeg3o8JsICdAKm38OhtGio2zveBd37Jm8IEWUmfAAAAAElFTkSuQmCC";
